+++ sql/core/src/main/scala/org/apache/spark/sql/execution/adaptive/AdaptiveSparkPlanHelper.scala
@@ -17,7 +17,6 @@
 
 package org.apache.spark.sql.execution.adaptive
 
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.execution.SparkPlan
 
 /**
@@ -135,25 +134,5 @@ trait AdaptiveSparkPlanHelper {
   def stripAQEPlan(p: SparkPlan): SparkPlan = p match {
     case a: AdaptiveSparkPlanExec => a.executedPlan
     case other => other
-  }
-
-  /**
-   * Returns a cloned [[SparkSession]] with all specified configurations disabled, or
-   * the original [[SparkSession]] if all configurations are already disabled.
-   */
-  def getOrCloneSessionWithConfigsOff(
-      session: SparkSession,
-      configurations: Seq[String]): SparkSession = {
-    val configsEnabled = configurations.filter(
-      session.sessionState.conf.getConfString(_, "false").toBoolean)
-    if (configsEnabled.isEmpty) {
-      session
-    } else {
-      val newSession = session.cloneSession()
-      configsEnabled.foreach(conf => {
-        newSession.sessionState.conf.setConfString(conf, "false")
-      })
-      newSession
-    }
   }
 }
